Struct std::os::unix::net::UnixDatagram

pub struct UnixDatagram(_);

This is supported on Unix only.

A Unix datagram socket.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let socket = UnixDatagram::bind("/path/to/my/socket")?;
    socket.send_to(b"hello world", "/path/to/other/socket")?;
    let mut buf = [0; 100];
    let (count, address) = socket.recv_from(&mut buf)?;
    println!("socket {:?} sent {:?}", address, &buf[..count]);
    Ok(())
}
Run

Implementations

impl UnixDatagram

pub fn bind<P: AsRef<Path>>(path: P) -> Result<UnixDatagram>

This is supported on Unix only.

Creates a Unix datagram socket bound to the given path.

Examples

use std::os::unix::net::UnixDatagram;

let sock = match UnixDatagram::bind("/path/to/the/socket") {
    Ok(sock) => sock,
    Err(e) => {
        println!("Couldn't bind: {:?}", e);
        return
    }
};
Run

pub fn unbound() -> Result<UnixDatagram>

This is supported on Unix only.

Creates a Unix Datagram socket which is not bound to any address.

Examples

use std::os::unix::net::UnixDatagram;

let sock = match UnixDatagram::unbound() {
    Ok(sock) => sock,
    Err(e) => {
        println!("Couldn't unbound: {:?}", e);
        return
    }
};
Run

pub fn pair() -> Result<(UnixDatagram, UnixDatagram)>

This is supported on Unix only.

Creates an unnamed pair of connected sockets.

Returns two UnixDatagramss which are connected to each other.

Examples

use std::os::unix::net::UnixDatagram;

let (sock1, sock2) = match UnixDatagram::pair() {
    Ok((sock1, sock2)) => (sock1, sock2),
    Err(e) => {
        println!("Couldn't unbound: {:?}", e);
        return
    }
};
Run

pub fn connect<P: AsRef<Path>>(&self, path: P) -> Result<()>

This is supported on Unix only.

Connects the socket to the specified address.

The send method may be used to send data to the specified address. recv and recv_from will only receive data from that address.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    match sock.connect("/path/to/the/socket") {
        Ok(sock) => sock,
        Err(e) => {
            println!("Couldn't connect: {:?}", e);
            return Err(e)
        }
    };
    Ok(())
}
Run

pub fn try_clone(&self) -> Result<UnixDatagram>

This is supported on Unix only.

Creates a new independently owned handle to the underlying socket.

The returned UnixDatagram is a reference to the same socket that this object references. Both handles can be used to accept incoming connections and options set on one side will affect the other.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::bind("/path/to/the/socket")?;
    let sock_copy = sock.try_clone().expect("try_clone failed");
    Ok(())
}
Run

pub fn local_addr(&self) -> Result<SocketAddr>

This is supported on Unix only.

Returns the address of this socket.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::bind("/path/to/the/socket")?;
    let addr = sock.local_addr().expect("Couldn't get local address");
    Ok(())
}
Run

pub fn peer_addr(&self) -> Result<SocketAddr>

This is supported on Unix only.

Returns the address of this socket's peer.

The connect method will connect the socket to a peer.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.connect("/path/to/the/socket")?;

    let addr = sock.peer_addr().expect("Couldn't get peer address");
    Ok(())
}
Run

pub fn recv_from(&self, buf: &mut [u8]) -> Result<(usize, SocketAddr)>

This is supported on Unix only.

Receives data from the socket.

On success, returns the number of bytes read and the address from whence the data came.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    let mut buf = vec![0; 10];
    let (size, sender) = sock.recv_from(buf.as_mut_slice())?;
    println!("received {} bytes from {:?}", size, sender);
    Ok(())
}
Run

pub fn recv(&self, buf: &mut [u8]) -> Result<usize>

This is supported on Unix only.

Receives data from the socket.

On success, returns the number of bytes read.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::bind("/path/to/the/socket")?;
    let mut buf = vec![0; 10];
    sock.recv(buf.as_mut_slice()).expect("recv function failed");
    Ok(())
}
Run

pub fn send_to<P: AsRef<Path>>(&self, buf: &[u8], path: P) -> Result<usize>

This is supported on Unix only.

Sends data on the socket to the specified address.

On success, returns the number of bytes written.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.send_to(b"omelette au fromage", "/some/sock").expect("send_to function failed");
    Ok(())
}
Run

pub fn send(&self, buf: &[u8]) -> Result<usize>

This is supported on Unix only.

Sends data on the socket to the socket's peer.

The peer address may be set by the connect method, and this method will return an error if the socket has not already been connected.

On success, returns the number of bytes written.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.connect("/some/sock").expect("Couldn't connect");
    sock.send(b"omelette au fromage").expect("send_to function failed");
    Ok(())
}
Run

pub fn set_read_timeout(&self, timeout: Option<Duration>) -> Result<()>

This is supported on Unix only.

Sets the read timeout for the socket.

If the provided value is None, then recv and recv_from calls will block indefinitely. An Err is returned if the zero Duration is passed to this method.

Examples

use std::os::unix::net::UnixDatagram;
use std::time::Duration;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.set_read_timeout(Some(Duration::new(1, 0)))
        .expect("set_read_timeout function failed");
    Ok(())
}
Run

An Err is returned if the zero Duration is passed to this method:

use std::io;
use std::os::unix::net::UnixDatagram;
use std::time::Duration;

fn main() -> std::io::Result<()> {
    let socket = UnixDatagram::unbound()?;
    let result = socket.set_read_timeout(Some(Duration::new(0, 0)));
    let err = result.unwrap_err();
    assert_eq!(err.kind(), io::ErrorKind::InvalidInput);
    Ok(())
}
Run

pub fn set_write_timeout(&self, timeout: Option<Duration>) -> Result<()>

This is supported on Unix only.

Sets the write timeout for the socket.

If the provided value is None, then send and send_to calls will block indefinitely. An Err is returned if the zero Duration is passed to this method.

Examples

use std::os::unix::net::UnixDatagram;
use std::time::Duration;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.set_write_timeout(Some(Duration::new(1, 0)))
        .expect("set_write_timeout function failed");
    Ok(())
}
Run

An Err is returned if the zero Duration is passed to this method:

use std::io;
use std::os::unix::net::UnixDatagram;
use std::time::Duration;

fn main() -> std::io::Result<()> {
    let socket = UnixDatagram::unbound()?;
    let result = socket.set_write_timeout(Some(Duration::new(0, 0)));
    let err = result.unwrap_err();
    assert_eq!(err.kind(), io::ErrorKind::InvalidInput);
    Ok(())
}
Run

pub fn read_timeout(&self) -> Result<Option<Duration>>

This is supported on Unix only.

Returns the read timeout of this socket.

Examples

use std::os::unix::net::UnixDatagram;
use std::time::Duration;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.set_read_timeout(Some(Duration::new(1, 0)))
        .expect("set_read_timeout function failed");
    assert_eq!(sock.read_timeout()?, Some(Duration::new(1, 0)));
    Ok(())
}
Run

pub fn write_timeout(&self) -> Result<Option<Duration>>

This is supported on Unix only.

Returns the write timeout of this socket.

Examples

use std::os::unix::net::UnixDatagram;
use std::time::Duration;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.set_write_timeout(Some(Duration::new(1, 0)))
        .expect("set_write_timeout function failed");
    assert_eq!(sock.write_timeout()?, Some(Duration::new(1, 0)));
    Ok(())
}
Run

pub fn set_nonblocking(&self, nonblocking: bool) -> Result<()>

This is supported on Unix only.

Moves the socket into or out of nonblocking mode.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.set_nonblocking(true).expect("set_nonblocking function failed");
    Ok(())
}
Run

pub fn take_error(&self) -> Result<Option<Error>>

This is supported on Unix only.

Returns the value of the SO_ERROR option.

Examples

use std::os::unix::net::UnixDatagram;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    if let Ok(Some(err)) = sock.take_error() {
        println!("Got error: {:?}", err);
    }
    Ok(())
}
Run

pub fn shutdown(&self, how: Shutdown) -> Result<()>

This is supported on Unix only.

Shut down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).

use std::os::unix::net::UnixDatagram;
use std::net::Shutdown;

fn main() -> std::io::Result<()> {
    let sock = UnixDatagram::unbound()?;
    sock.shutdown(Shutdown::Both).expect("shutdown function failed");
    Ok(())
}
Run